0.2.0
updated generate js to use switch statmenets to verify input value and act accordingly then imported function and called inside of the readme file with the parameter of license value from input. now a badge is rendered and a link is provided and a preview section all in relation to specified license type.